fix/improve kumascript tests
