+++ kumascript/src/api/wiki.js
@@ -103,7 +103,7 @@ module.exports = {
   },
 
   // Returns the page object for the specified page.
-  async getPage(path) {
+  getPage(path) {
     this.info.getPage(path || this.env.url);
   },
 
@@ -124,13 +124,13 @@ module.exports = {
   // Special note: If ordered is true, pages whose locale differ from
   // the current page's locale are omitted, to work around misplaced
   // localizations showing up in navigation.
-  async tree(path, depth, self, reverse, ordered) {
+  tree(path, depth, self, reverse, ordered) {
     // If the path ends with a slash, remove it.
     if (path.substr(-1, 1) === "/") {
       path = path.slice(0, -1);
     }
 
-    var pages = await this.page.subpages(path, depth, self);
+    var pages = this.page.subpages(path, depth, self);
 
     if (reverse == 0) {
       pages.sort(alphanumForward);
