--- kumascript/tests/macros/fixtures/subpages-depth-of-0.json
@@ -1,7 +1,0 @@
-{
-  "locale": "en-US",
-  "url": "/en-US/docs/Web/HTTP/Basics_of_HTTP",
-  "subpages": [],
-  "slug": "Web/HTTP/Basics_of_HTTP",
-  "title": "Basics of HTTP"
-}
