--- kumascript/tests/macros/fixtures/translations.json
@@ -1,105 +1,0 @@
-{
-  "json_modified": "2017-07-02T21:56:43.462560",
-  "uuid": "579a7d0a-8b01-4257-a067-3a0098eb0798",
-  "title": "Basics of HTTP",
-  "url": "/en-US/docs/Web/HTTP/Basics_of_HTTP",
-  "tags": ["HTTP", "Overview"],
-  "translations": [
-    {
-      "uuid": "579a7d0a-8b01-4257-a067-3a0098eb0798",
-      "title": "Conceptos básicos de HTTP",
-      "url": "/es/docs/Web/HTTP/Basics_of_HTTP",
-      "tags": ["TopicStub", "HTTP", "Overview", "NeedsTranslation"],
-      "summary": "El protocolo HTTP es un protocolo ampliable, es decir se puede añadir \"vocabulario\". HTTP está basado en unos pocos conceptos básicos como el concepto de recursos y URIs, una estructura sencilla de mensajes, y una arquitectura de cliente-servidor para ordenar el flujo de las comunicaciones. A demás de estos conceptos, a lo largo de su desarrollo han aparecido otros nuevos y se han añadido funcionalidades y reglas semánticas, creando nuevos métodos y cabeceras.",
-      "localization_tags": ["inprogress"],
-      "locale": "es",
-      "last_edit": "2017-07-02T21:56:37.330543",
-      "review_tags": []
-    },
-    {
-      "uuid": "579a7d0a-8b01-4257-a067-3a0098eb0798",
-      "title": "Basics of HTTP",
-      "url": "/fr/docs/Web/HTTP/Basics_of_HTTP",
-      "tags": ["TopicStub", "HTTP", "Overview", "NeedsTranslation"],
-      "summary": "HTTP is a pretty extensible protocol. It relies on a few basics concepts like the notion of resources and URIs, a simple structure of messages, and a client-server structure for the communication flow. On top of these basics concepts, numerous extensions have appeared over the years, adding new functionality and new semantics by creating new HTTP methods or headers.",
-      "localization_tags": ["inprogress"],
-      "locale": "fr",
-      "last_edit": "2016-09-08T20:43:34",
-      "review_tags": []
-    },
-    {
-      "uuid": "579a7d0a-8b01-4257-a067-3a0098eb0798",
-      "title": "Basics of HTTP",
-      "url": "/ja/docs/Web/HTTP/Basics_of_HTTP",
-      "tags": ["TopicStub", "HTTP", "Overview", "NeedsTranslation"],
-      "summary": "HTTP is a pretty extensible protocol. It relies on a few basics concepts like the notion of resources and URIs, a simple structure of messages, and a client-server structure for the communication flow. On top of these basics concepts, numerous extensions have appeared over the years, adding new functionality and new semantics by creating new HTTP methods or headers.",
-      "localization_tags": ["inprogress"],
-      "locale": "ja",
-      "last_edit": "2016-09-08T20:43:34",
-      "review_tags": []
-    },
-    {
-      "uuid": "579a7d0a-8b01-4257-a067-3a0098eb0798",
-      "title": "Basics of HTTP",
-      "url": "/ko/docs/Web/HTTP/Basics_of_HTTP",
-      "tags": ["TopicStub", "HTTP", "Overview", "NeedsTranslation"],
-      "summary": "HTTP is a pretty extensible protocol. It relies on a few basics concepts like the notion of resources and URIs, a simple structure of messages, and a client-server structure for the communication flow. On top of these basics concepts, numerous extensions have appeared over the years, adding new functionality and new semantics by creating new HTTP methods or headers.",
-      "localization_tags": ["inprogress"],
-      "locale": "ko",
-      "last_edit": "2016-08-12T04:29:34",
-      "review_tags": []
-    },
-    {
-      "uuid": "4c1568bd-c5a6-451e-bd6d-ddafcebfa4f6",
-      "title": "Básico sobre HTTP",
-      "url": "/pt-BR/docs/Web/HTTP/Basico_sobre_HTTP",
-      "tags": [],
-      "summary": "HTTP é um protocolo bem extensivo. Isso depende um pouco do conceito básico com noção de recursos e URIs, uma simples estrutura de mensagens, e uma estrutura de cliente-servidor para a comunicação ocorrer. Em cima destes conceitos básicos, várias versões surgiram ao longo do tempo, adicionando novas funcionalidades e novas semanticas para criar novo metohos HTTP ou cabeçalhos. the most adequate response",
-      "localization_tags": [],
-      "locale": "pt-BR",
-      "last_edit": "2017-03-08T21:53:54.258318",
-      "review_tags": []
-    },
-    {
-      "uuid": "579a7d0a-8b01-4257-a067-3a0098eb0798",
-      "title": "Basics of HTTP",
-      "url": "/ru/docs/Web/HTTP/Basics_of_HTTP",
-      "tags": ["TopicStub", "HTTP", "Overview", "NeedsTranslation"],
-      "summary": "HTTP is a pretty extensible protocol. It relies on a few basics concepts like the notion of resources and URIs, a simple structure of messages, and a client-server structure for the communication flow. On top of these basics concepts, numerous extensions have appeared over the years, adding new functionality and new semantics by creating new HTTP methods or headers.",
-      "localization_tags": ["inprogress"],
-      "locale": "ru",
-      "last_edit": "2016-09-08T20:43:34",
-      "review_tags": []
-    },
-    {
-      "uuid": "f2b4286e-4db3-476f-948d-96517ad98da0",
-      "title": "HTTP 基础",
-      "url": "/zh-CN/docs/Web/HTTP/Basics_of_HTTP",
-      "tags": ["概览", "HTTP"],
-      "summary": "HTTP 是一个拓展性非常好的协议. 它构建在以下基础之上: 一些像资源或是 URI 这样的基本概念, 一个简单的消息结构, 一个客户端-服务器结构的通信流. 在这些基础概念之上, 近年来已经出现了许多拓展, 以增加新的 HTTP 方法或首部的方式为 HTTP 协议增加了新的功能和语义.",
-      "localization_tags": ["inprogress"],
-      "locale": "zh-CN",
-      "last_edit": "2017-04-26T20:04:37.343007",
-      "review_tags": ["technical", "editorial"]
-    }
-  ],
-  "modified": "2016-12-13T09:12:47.420798",
-  "label": "Basics of HTTP",
-  "localization_tags": [],
-  "locale": "en-US",
-  "id": 189027,
-  "last_edit": "2016-09-08T20:43:34",
-  "summary": "HTTP is a pretty extensible protocol. It relies on a few basics concepts like the notion of resources and URIs, a simple structure of messages, and a client-server structure for the communication flow. On top of these basics concepts, numerous extensions have appeared over the years, adding new functionality and new semantics by creating new HTTP methods or headers.",
-  "sections": [
-    {
-      "id": "Quick_Links",
-      "title": null
-    },
-    {
-      "id": "Articles",
-      "title": "Articles"
-    }
-  ],
-  "slug": "Web/HTTP/Basics_of_HTTP",
-  "review_tags": []
-}
